✨ feat: 커맨드 Input 컴포넌트 분리 및 옵션 핸들링 추가
