+++ apps/web/src/chatbot/index.ts
@@ -8,7 +8,9 @@ export interface ChatbotFunction {
   name: string;
   type: 'API_QUERY' | 'API_CONFIG' | 'WIZBOT_CONFIG';
   optionLabel?: string | null;
-  optionInput?: (userId: number) => { type: 'text' } | { type: 'select'; options: string[] };
+  optionInput?: (
+    userId: number,
+  ) => Promise<{ type: 'text' } | { type: 'select'; options: { key: string; value: string }[] }>;
   description: JSX.Element;
   descriptionShort: string;
   usage: (command: string, option?: string) => JSX.Element;
